Atgr 28 (#25)
* Add Identity endpoints

* Add register endpoint

Enable authentication

* Fix compilation error

* Add SignIn and SignUp page

Store tokens in store and session storage

* ATGR-28: Log in
- Added link to register page

+++ AnytimeGear/AnytimeGear.Server/Dtos/ProductListResponseDto.cs
@@ -4,8 +4,8 @@ public class ProductListResponseDto
 {
     public ICollection<ProductResponseDto> Items { get; set; } = [];
     public int TotalCount { get; set; }
-    public int MinPrice { get; set; }
-    public int MaxPrice { get; set; }
+    public decimal MinPrice { get; set; }
+    public decimal MaxPrice { get; set; }
     public ICollection<ProductBrandDto> Brands { get; set; } = [];
     public required string SortKey { get; set; }
     public required string SortOrder { get; set; }
